pyproject.toml
[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[project]
name = "plugin-examples"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "jsonschema>=4.21",
    "pyyaml>=6.0",
    "requests>=2.31",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-timeout>=2.3",
]

[project.scripts]
plugin-examples = "plugin_examples.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
timeout = 30
